Rework type functionality in info
We want to be able to support flags for particles, so we rearrange the
meaning of the type field found in the particle info structure. Instead
of coding the particle fluid number, with negative values for the
boundaries or moving objects, we split it into two bytes, with the lower
byte split in two nibbles. The lowest nibble (bits 0..3) codes the fluid
number for multi-fluid, the upper lower nibble (bits 4..7) code the
moving object, the upper byte (bites 8..15) are reserved for flags.
No flags are defined yet.

The FLUID() and NOT_FLUID() macros are adjusted to fit the new usage,
and no other change in the code is required.