Fix thread synchronization in partial reduction
Move the __syncthreads in dtataptBlockReduce to before the shmem
operation, to ensure that all warps in the block have committed their
data when the reduction starts.

There is no need for a 'final' __syncthreads() because the loop is
exited when only the first thread did something, and it's still the
first thread that does the final data save.