Always reinit the gage height
The previous gage value was being used as starting point when averaging
the positions of neighboring particles, resulting in the gage position
to be slightly shifted towards the old value. This is particularly
visible when looking at the initial vertical fluid collapse.

Fix by setting it to 0 before the sum loop.