Add procedure for initialization of gamma coefficient
Initialization of gamma is implemented in initializeGammaAndGradGamma()
function. Firstly, the gradient of gamma is calculated for every parti-
cle using simplified approximation. Then particles with non-zero values
of gamma gradient are moved away from the solid walls. Using an itera-
tive procedure fluid particles are returned on their initial positions.
Gradient of gamma is recalculated on every time step and gamma value is
updated according to the evolution equation, starting from the value of
1.0.