Make hash key size configurable
By default we keep the 32-bit hash key that matches the gridHash, but we
allow its size to be enlarged to include additional information. The
gridHash will always occupy the most significant bits of the particle
hash. Preliminary code for 32 and 64 bit hash keys is included, with the
64-bit hash key just a shifted version of th 32-bit hash key for the
time being.