Update continuity and momentum equations according to new boundary model
Both equations are modified only for formulation SPH_F1. In the momentum
equation only pressure term is corrected at the boundary, viscosity term
for the new boundary will be implemented later.

In addition, some corrections have been introduced to the initialization
of density. The density value at the boundary element is calculated as
an average of the values at three vertices of this element. Density and
pressure at the vertex particles represent hydrostatic pressure profile.