Add Ferrari correction to continuity equation
Ferrari correction is a volume diffusion term introduced in continuity
equation. The implemented formulation corresponds to a paper of
A.Mayrhofer et al. from Spheric 7th (see Eqs.46-47).

The usage is defined by the parameter "m_simparams.ferrari", which value
is between 0 and 1. When it is zero, no correction is applied.