Fix zero mass of boundary particles
We are using Crixus preprocessor to create input files with arbitrary
geometry. It zeroes masses (actually volumes) of boundary particles.
For the calculations it is not a problem, since these masses are not
used in any equation. By the way it will cause problems together with
a new feature proposed by Giuseppe Bilotta for outlet - specifying of
particles, which are out of outlet, with zero mass.

Currently if compiled with InputProblem masses of boundary particles
are set to 99.