calcGridHash: max before min
Since gridSize is unsigned the min() will be an unsigned comparison.
When gridPos is negative (e.g. -1), this would turn it into a very large
positive number and gridSize - 1 would be used instead of gridPos in the
subsequent max() call. As a result, particles with position lower than
m_origin would be hashed as being on the opposite side of the domain,
and not be included in the proper neighbor lists.

By first computing the max(0, gridPos), we ensure that the result is
non-negative, and that the subsequent min() will not cause
signed/unsigned problems. This ensures that the particles' grid
positions are clamped correctly, ensuring the correctness of the hash.