Static pointer to GlobalData for the signal handlers
A static pointer (gdata_static_pointer) to the instance of GlobalData
allocated in the main is declare in GlobalData.h. This aims to make
variables such as quit_request and save_request accessible by the
signal handlers, which can not have any visibility of the GlobalData
instance allocated inside the main().