WriterType moved to GlobalData.h + workaround
WriterType is now declared in GlobalData.h. This is not only more
logical but also useful because using Problem::WriterType in the
GlobalData struct is difficult with the current forward declaration.
Unfortunately, since the old main (PS, etc.) still uses the one
declared in Problem, the GPUSPH::createWriter() method is currently
(and temporarily) quite awkward since it deals with both.