Also define round_up, use for the neibslistSize computation
Sadly, due to C++ not being smart enough with type promotion and
template parameters, NEIBINDEX_INTERLEAVE has to be an uint to be of the
same type as m_numParticles (other alternatives would have been forcing
a template specialization when using round_up or casting
NEIBINDEX_INTERLEAVE as uint when passing it as an argument; and then
people complain because _macros_ are bad? sheesh).