Get rid of m_numPartsFmax
There is no need to store numPartsFmax and pass it to the forces()
function at each invocation. Its number is always exactly the number of
blocks in the forces kernel call, so we can just recycle that number.

This has the advantage of being more flexible with respect to the forces
kernel being run on a variable number of particles (of course, the cfl
array should still be allocated for the maximum number of elements it
might ever contain).

While we're at it, fix the amount of memory which we claim is allocated
(the temporary cfl array for reductions is not being considered).