Quick & dirty trick to avoid zero dt if a device has no fluid
When a device runs out of fluid particles (plausible without load
balancing and with a trivial split, e.g. horizontal cut for
DamBreak3D) the retrieved dt is zero. We workaround this problem
by simply not taking into consideration zero dts. This must
be considered a temporary fix until a more robust solution is
implemented.