Update k and e following the existing predictor-corrector time scheme.
Values of k and e are updated in the "euler" kernel according to the
following time scheme:
Forces(Predictor):
  dk/dt(pos(n), vel(n)) and de/dt(pos(n), vel(n))
Euler(Predictor):
  k(n+1/2) = k(n) + dk/dt(pos(n), vel(n))*dt/2
  e(n+1/2) = e(n) + de/dt(pos(n), vel(n))*dt/2
Forces(Corrector):
  dk/dt(pos(n+1/2), vel(n+1/2)) and de/dt(pos(n+1/2), vel(n+1/2))
Euler(Corrector):
  k(n+1) = k(n) + dk/dt(pos(n+1/2), vel(n+1/2))*dt
  e(n+1) = e(n) + de/dt(pos(n+1/2), vel(n+1/2))*dt

Values of dk/dt and de/dt are stored in float2 array dkde.

Currently, eddy viscosity is also stored in an individual array, and
its value is computed from k and e in the "forces" kernel.