Fix: passing the range to the kernels as well as the numParts
Previous to this commit, when a kernel needed to be executed only
on a range of particles, the number of internal particles was
passed in place of numParticles. However, this led to incorrect
texture binding, since the external particles were left out of
the texture and thus not read. Now also numParticles is passed
to the buildneibs, forces and euler kernels for correct texture
binding. Moreover, m_particleRangeEnd is used instead of the
number of internal particles, since we plan to run kernels on
arbitrary subranges of particles (for striping).
ParticleSystem is also updated for correct compiling.