Extend equations for dk/dt and de/dt with diffusive terms.
In "forces" kernel expressions for computing dk/dt and de/dt is
extended with diffusive terms, according to eqns (A.3) and (A.9) from
Ferrand et al.(2012). For de/dt only first term of (A.9) is implemen-
ted. The second term, which includes interaction with boundary elements
will be rewritten basing on the other boundary conditions on e.