Add boundary conditions on k and e
Values of k at the vertex particles are calculated by weighted sum-
mation of k over the neighborig fluid particles (eqn. A.4):
  k(v) = 1/alpha(v) * SUM[mass(b)/rho(b)*k(b)*W(bv)],
where alpha(v) is the same as used for computing values of density
and pressure at the vertex particles.

The rate of dissipation of turbulent energy (e) at the vertex particle
is calculated by the following formula:
  e(v) = C_mu^(3/4) * k(v)^(3/2) / (K * deltap),
where C_mu = 0.09 and K = 0.41 are the constants.

After that, values of k and e at the boundary elements are simply
computed by averaging of the values of three vertex particles be-
longing to this boundary element.