Add boundary term to diffusive part of the equation for de/dt
Diffusive part of equation for de/dt is extended with a boundary
term. The actual boundary term is different from the one written
in equation (A.9). Now the full expression for diffusive term reads
as:
 (A.9) = 1/gamma_a * SUM{fl. & vert. parts}[...] - 1/(gamma_a*rho_a) *

 * SUM{bound.elems} [
      (mu_0_a + mu_t_a/sigma_e)(-k_a^(3/2)*C_mu^(3/4) / (K*dr_as^2) +
    + (mu_0_s + mu_t_s/sigma_e)(-k_s^(3/2)*C_mu^(3/4) / (K*dp^2) *
    * |gradGamma_as|
   ],

where C_mu = 0.09, sigma_e = 1.3, and K = 0.41 are the constants,
dr_as is defined as maximum between dp and distance from the particle
a to boundary element s along the normal.