Add "slip" boundary condition for k-e model
When k-e model is turned on, velocity at the vertex particles is used
to impose "slip" boundary condition. This velocity is not used to move
the vertex particles.
Starting from zero value, velocity is updated according to eq.(A.2)
from Ferrand et al. (2012) paper. Velocity is updated inside "euler"
kernel.
Currently only first term of eq.(A.2), which is a viscous term, is
implemented. That means that with some delay velocity at the wall is
the same as velocity of the fluid neighbours.