Disabled --fast-math in Makefile
With fm enabled, CPU and GPU can put the same particle in different
cells. This is a problem in mgpu since the receiving device does not
recognize a particle and drops it.
It can be enabled again after the computation of the cell containing
a particle is being implemeted on the GPU, even before the host sort.