Compute velocity at the boundary elements by averaging over vertices
Velocity at every boundary element is calculated as an overage over
velocities of three vertices. This is a virtual velocity used to impose
"slip" boundary conditions, when k-e model is turned on. For other
viscosity models the average of zero velocities (which is consequently
zero) will be calculated, imposing "no-slip" boundary conditions.