Fix: passing correct maxParticles to euler
When a particle is created, the new id is (oldId + allocatedParticles).
However, the device allocated particles was passed instead of the
host one, causing pairs of "clone" particles.
This will work for multinode as well, as long as allocatedParticles
is multiplied by dev_global_index (reducing the maximum number of
particles).