Script to estimate deltap to get a given number of particles
The script works by extrapolating from two "data points" (particles in a
problem at given values of deltap) and then solving a cubic equation to
find the deltap to be used to get a problem with a given number of
particles.

The formula is not perfect, but it's a good start (also, rather than a
script, it should be integrated into Problem itself).