Fix in handling empty devices: resetting cellStarts
Need to set cells as empty if a device has no particles, or
the first dump of cellStarts will produce uninitialized values
(which will be used in importPeerCells). Now setDeviceCellsAsEmpty()
is called instead of explicitly resetting with cudaMemset() in the
reorder kernel wrapper, but it is called anyway despite the number
of particles.