Fix cell pos to global pos formulas
A few formulas all around assumed that the wolrd origin was null. While
the use of homogeneous precision does make shifting the world origin
less important, it can still gain us 1 bit of precision (even if it's
just in double), so keep supporting it correctly.

Notice that in ParticleSystem we use the built-in single-precision
worldOrigin when possible, but use the Problem's double-precision one at
write time.