Added --num-hosts option to allow multiple processes per host
If --num-hosts INT is specified, the simulator is aware that the
gdata->mpi_nodes processes will be distributed across the given
number of hosts and thus computes an appropriate offset to be
added to the device number. This assumes round robin slot scheduling
is being used by the underlying MPI implementation.