Check that dt is not too small
Dt is compared to the machine epsilon (FLT_EPSILON) and then with t
itself (if t is much bigger, the sum might be subject to cancellation).

If dt is too small, a quit request is issued, since in most cases
dt does not recover quickly and, in general, this might be signal
of another problem.