Adding new buffer VERTPOS which contains vertex positions relative to a segment.
In detail VERTPOS is an array of 3 float2. It is only set for segment s. The position is given in the plane \uvec{n}_s. The first coordinate is equal to \uvec{e}_i (unit coordinate in direction i) where i is equal to the coordinate for which n_{s,i} is minimal. The second coordinate is equal to \uvec{n}_s \times \uvec{e}_i. This avoids having to know neighbours of neighbours which can be tricky, particularly in multigpu.
TODO: needs to be allocated only for segments and not for all particles.