CELLTYPE_BITMASK defined differently according to HASH_KEY_SIZE
If HASH_KEY_SIZE is 64 bit wide, any bitwise AND with CELLTYPE_BITMASK
reset the first two bits of a word; otherwise, it has no effect.
CELLTYPE_BITMASK is anyway 32 bits wide.

Together with GRIDHASH_BITSHIFT, this helps making bitwise hash
transformations with the same constant, regardless the length
of the hashKey.