forces kernel: swap info selector and neib loop
Iterate over all neighbors regardless of the particle type for
the current particle, and only operate conditionally on it when it comes
to doing the actual particle/particle interaction.

While this presently wastes some registers in some forces kernel
specializations, it improves the logic allowing for the upcoming more
aggressive cleanups.