Have a define for particle types
Some upcoming forces kernel changes will rely on the particle type. This
could already be extracted with appropriate macros, but there wasn't an
easy way to actually check its value and compare it against the known
types.

We now have an enum ParticleType, and the previous defines are now based
on the enum values.