New mechanism for importing bursts of cells from other nodes
CPU profiling showed that iterating on the cells to compute the bursts
was expensive, esp. when importing network cells, which required reading
the neighbors of each cell. To minimize its impact on the performance,
the bursts are now computed once in the method computeCellBursts();
the list of bursts is stored by each worker in m_bursts vector.

importNetworkPeerEdgeCells() has been "split": now it calls
transferBurstsSizes() when appeding and transferBursts() in both
append and update. The former exchanges the cell sizes with other
nodes while the latter sends the particles contained in those cells.

Finally, device copies of cellStart and cellEnd are updated with a single
transfer rather than using one memcpy per cell.

Performance on cluster was increased almost by a factor 10 after
these optimizazions.