Particle range cached in CellBurst
Burst edges in terms of cells change when the device map change, but
their edges in terms of particle indices must be recomputed after
every buildneibs. These were computed on the fly in transferBursts().

CellBurst now has also firstParticle and lastParticle fields, which
are used to cache che burst interval in terms of particle indices.
Now particle ranges are computed in transferBurstsSizes() (so once
per buildneibs) and stored in the new fields.

Also, the detection of the cell range had a minor improvement.

A quick test on 4 nodes showed a performance improvement of ~4%
(in MIPPS).