Fix: do not recompute particle hash at iteration 0
Particles are filled in double precision by the Problem and then
translated into cell + relative_float coordinates. The two coordinate
systems might slightly disagree on which cell a particle belongs to.

In general this is not a problem since calcHashDevice() can updates the
cellHash and the relative coordinates. However, if this cell change
happens between two devices (i.e. the cell change is also a device change)
and at iteration 0 (i.e. when devices do not have yet a copy of the edge
particles of neighbording devices) ,then the particles are lost and a
warning is printed (note: in branches featuring the outlets, no warning
is printed if there is an outlet in the problem).

In other words, if the GPU and the CPU disagree on which device should
own a particle, that particle is lost before a copy is acquired by the
recipient device. To reproduce the problem with DamBreak3D:

1. Use fillDeviceMapByAxis(Z_AXIS) in fillDeviceMap()
2. Run a 2-devices simulation with deltap 0.005

To avoid this, we tell calcHashDevice() to use the CPU-computed
particle hash (and thus set only the particleIndex) if 1. the
simulation is multi-device, and 2. iterations == 0. The
condition is passed to the kernel by setting the particleHash array
pointer to NULL.