Improving fix of commit 4c6bfdb5
First calcHash does not change the particle hash computed on host.
However, hashes computed on host have clean high bits, i.e. cells
are not marked as internal/edge/external.

This is not an issue until bursts are compacted solely on the basis
of the linear cell index. Since edge cells are compacted at the end
of the particle indices (if high bits are set), we could use fewer,
longer bursts in some configurations (e.g. 2 devices, linearization
orthogonal to split axis). If we want to compact bursts also on the
basis of edge cells, then we want high bits to be set already at
iteration 0.

Instead of adding conditionals to calcHash, another kernel is
implemented (fixHash) which only sets particleIndex and the high
bits of the particleHash.

Other options could have been:
- Compute hashes with high bits alread set on host (not efficient)
- Adding a bool to the signiture of calcHash
- Using posArray==NULL as condition (like particleHash is used in
  commit 4c6bfdb5
- Use more bursts for iteration 0