More compact bursts with bursts as lists of cells, not cell ranges
Cell bursts are now lists of cells, not cell ranges (thus, linearized
indices are not necessarily consecutive anymore). This allows for
more compact bursts in some circumstances, esp. when the linearizazion
is not optimal: for example, with XYZ linearization and Y split outgoing
cells are often consecutive in memory but not in the cell indices.

It is possible to temporarily disable this optimization (for debug or
profiling purposes) by commenting one line (see usage of
"edging" bool).

After this commit, performance for a DamBreak3D with 873k particles
increases by about 5%.

This is possible only after commit 19b9b828, otherwise at iteration
0 outgoing cells would be sparse and the wrong data would be transferred.