New --striping command line options; striping disabled by default
Transferring the edging cells in bursts is pretty efficient, in
particular when simulating with 2 devices only, possibly both in
the same PCI domain (in this scenario, only one transfer is needed,
and this is two orders of magnitude faster than forces kernel).
This is also the most common multi-GPU usage, so striping is
disabled by default.
In the other cases (e.g. 3+ devices or multi-node simulations) it
is possible to enable it with --striping.