Fix: explicit host staging should use the same stream for D->H and H->D
If different streams are used, transfers are theorically independent and
may overlap. Instead, we want the upload to wait for the download to be
over, or obsolete/null values could be read.
Now the same stream (m_asyncPeerCopiesStream) is used for both memcpys.
We still want cudaMemcpyAsync() so that transfers are asynchronous with
respect to the striped forces kernel.