New option --no-gpudirect to disable GPUDirect
GPUDirect support for RDMA allows passing a device pointer to MPI
calls, avoiding double host staging. This option disables the feature,
allowing multinode simulations with MPI libraries that are not
CUDA-aware.
It is also useful for CUDA-aware MPI libraries that do not allow
selecting the CUDA device after MPI_Init() (such as MVAPICH < 2.0).