Inverted --no-gpudirect (now: --gpudirect), disabled by default
GPUDirect is only available in some MPI implementations, so it is safer
to disable it.
But there is more: we need to investigate if GPUDirect is available
in multi-GPU multinode setups. It seems that even a recent, CUDA-aware
MPI implementation such as MVAPICH 2.0rc1 does not handle multiple
CUDA contexts and thus fails to register some CUDA events.
GPUDirect is currently recommended only for single-GPU, multinode
simulations.