Updating the h5sph reader and removing hardlinks to local files
The h5sph reader has been updated to be more streamlined such that its
inclusion in a problem is easier, hiding most work in HDF5SphReader,
such as the buffer allocation. The current problems in InputProblem had
their input files moved to the meshes folder so we no longer need
absolut paths that depend on the machine.