Fixing a gamma issue related to the larger epsilon
As the epsilon was quite large now there was a problem with the
computation of the vertex particle gamma. To fix this issue epsilon was
reset back to something close to numerical epsilon. The trigger for the
gamma smoothing is now set to slength/3 which is close to dp/2 which it
was before. This is better as it is now independent of dp and only
depends on h as it should be.