Fixing several issues in the gamma singularity fix
There were some problems associated with the use of non-dimensional
distances. Furthermore we use slength/2 instead of /3 now as this is
more stable. The averaging process using the sph smoothing length as
been replaced with one based on vertex particles only which also use a
1/x kernel for the averaging which ensures a kronecker delta property
which will be important for I/O boundary conditions.