Putting KENT into info.object
When reading a H5sph file the KENT buffer defines different types of
special boundaries. For now it will only be used for in/outflow but
later it could be used as well for floating objects. KENT can be set
using the *_sbgrid_* files used by Crixus and it is used as mere marker.
In the end GPUSPH needs to define what to do with each different
KENT/obj group.