Segments now have neighbours too
This was initially not required as segments would get their boundary
condtion from the vertices but this is no longer true. In fact the
neumann conditions are now enforced on segments and the dirichlet on
vertices. The interaction between the two happens with local averages.
This is in line with the finite volume approach where these two
different type of bcs are enforced on different elements as well.