Problem dependent imposition of open boundary conditions
This is a rather big commit but it also changes quite a few things:
1.) In InputProblem we now use preprocessor directives instead of the
string for specific problems
2.) The inflow/outflow particles are now identified via flags in the
particle type short. There is identification of velocity/pressure
imposition, tangential velocity prescription and whether the particle is
part of open boundaries. This removes the necessity to check for a
specific object number
3.) It is now possible to have a dedicated kernel for each problem that
imposes the velocity/pressure on open boundaries. The current
implementation can be seen in InputProblem_IO.cu. Note that this is
rather ugly but CUDA doesn't allow it to be any prettier, so to say it
with Linus:
http://www.journaldugeek.com/files/2012/06/nvidia-fu-600x3261.png