New ID computation for open boundaries
This uses an iteration counter that keeps track of the number of
iterations when particles were created. Plus the number of initial
non-fluid and vertex particles. It ensures that the IDs are unique even
though quite a few IDs are wasted. There is also an overflow check that
ends the simulation if the IDs overflow UINT_MAX.