New eulerVel to ensure tangential velocities only for fixed particles.
When keps is chosen as viscosity model, then the fixed boundary
particles would have an Eulerian velocity in order to balance friction
forces. In order to satisfy u.n = 0 for boundaries the gradient of gamma
was chosen as normal of the wall in order to make sure that this
condition is met for the Eulerian velocity. However, close to a corner
of an open boundary gradient of gamma is normally not equal to the
normal of the fixed walls. With this commit it is ensured that only the
segments associated to a vertex that belong to a fixed wall are used to
compute the vertex normal. This significantly stabilizes the flow and
removes spurious effects at wall / open boundary interfaces.