New forced moving objects
The previous approach was not consistent with how floating objects are
imposed. This new approach is now pretty much identical to that and
there is a new problem specific routine that allows the imposition of
the center of gravity, the translation and the rotation matrix of an
object to allow for arbitrary complex movement.