Fix in Gamma(), II approach: set minlRas only if reasonably close to a wall
This reverts previous commit (c3af2efb) and implements a different fix.
The fix consists again in setting minlRas only if reasonably close to a
wall, but with a simpler conditional.
Simulations appear stable now. With previous approach, running
CompleteSaExample with (box_density = 0.5 * water_density) causes a few
explosions, one of which (2.37s) so strong that the cube is thrown out
of the water.