Enabling basic ODE objects, + initializeODE(), cleanupODE(), initialize()
Problem::initialize() is virtual and empty by default (for compatibility
with previous Problem interface). It allocates ODE with the number of
requested objects and (will) resize the world size automatically. It is
called by GPUSPH at the beginning of the overall initialization.

Still missing: methods to set the mass. Now setting half fluid density.

Note about ODE bodies counters:

- m_simparams.numODEbodies is set by the user in the Problem constructor;
  in XProblem it is instead computed
- Problem::m_total_ODE_bodies is incremented in Problem::add_ODE_body()
- XProblem::m_numRigidBodies is a progressive counter incremented every
  time the user adds a body with dynamics; at the end of the constructor
  and of the initialize() method, this should be equal to
  m_total_ODE_bodies

Of course we will get rid of some of this stuff when XProblem will collapse
on Problem.