Add and call Object::updateODERotMatrix() to update ODE rotation matrix
This should not be called on planes. Unfortunately, it is not trivial
to make the call skip automatically if Object is a Plane, since ODE
does not offer any method to check if an object is placeable.
Possible alternative: make the method (non-pure) virtual and override
with warning in Plane class.