STLMesh: add m_origin member
m_origin represents the position of the origin of the mesh coordinates
in gpusph coordinates.

Leave to zero to make the mesh load in gpusph with the originary absolute
coordinates; move it to minimum of bounding box to make the lower corner
of the mesh bbox align with world origin.