BufferList: expose map-like interface without deriving from map
This allows us to abstract the implementation from the interface, which
is necessary to move double-buffering from the individual Buffer to the
BufferList level.

Throw a couple of const-correctnest fixes in, and prevent HotFile from
copying our BufferList (which leads to an actual deletion of our
buffers after the first write, since the destructor is now properly
called —it wasn't while BufferList subclassed map).