Refactor double-buffer handling
The number of arrays in a buffer now indicate the number of arrays
needed to represent the data. The need for double-buffering is instead
handled through a BufferAllocPolicy that in general would depends on
things such as the integration scheme used and possibly the SPH
formulation.

Each buffer now only knows about a single copy, and multiple copies are
handled through a MultiBufferList that takes care of all the
doule-buffering management.