Init s_hRbFirstIndex in XProblem instead of rbfirstindex in GPUWorker
This allows for setting a "personalized" offset for each object, which
simply consists in the lowest id of the particles of the object.
This allows for a more robust addressing in rbforces array and it works
also if objects are not filled consecutively.

The downside is that GlobalData is not const anymore for Problem. There
could be other ways to do it, but anyway we will split GlobalData in
subarrays for a cleaner structure.

The occasion was taken for a minor restructuration in copy_to_array.