Temporary fix to object offset handling, wait for merge with deswitch
s_hRbFirstIndex[] is used inside the forces kernel and therefore must be
indexed with the (unshifted) object id, since the kernel has no access
to the map obj_id -> ode_obj_id (which is also a generic rigid_body_id).

s_hRbLastIndex[], instead, is accessed in reduceRbForces(), which only
needs to iterate on rigid bodies (currently, ODE objects). Here holds
the opposite: here we iterate on ODE indices but we cannot access the
generic object_id (in the address space including open boundaries).

This difference causes wrong results in problems using both rigid bodies
and open boundaries (which have an object id but do not reduce forces).

The temporary fix is to fill the two arrays with different indices
already in XProblem::copy_to_array(). A cleaner fix would consist in:

- Renaming both arrays, since they refer to different address spaces and
  since the names were not descriptive anyway ("first" is actually an
  offset, "last" is the index of the reduction output);
- Separating in a cleaner way the list of objects (floating, moving and
  floating) from the list of objects requiring reduction (floating and
  moving).

The clean fix is delayed since we will merge with deswitch branch, which
implements a different solution.