Fix hash computation on multi-GPU
The particle refactoring (5306b2c13b11b44a3a5aef93642c151eb68761eb)
cleaned up the detection of FLUID and MOVING particles in the hash
calculation/fixing kernels (compared to the old, _wrong_, MOVINGNOTFLUID
macro).

With the old, wrong, macro, the kernels ran effectively for all
particles. With the fixed checks, kernels only ran on FLUID and MOVING,
as intended.

However, this caused an issue in multi-GPU simulations, since the
non-moving boundaries would end up not having the compactDeviceMap
OR-ed in their hash.

The fix is to run fixHashDevice() on all particles, and only skip
non-moving non-fluid particles in calcHashDevice() in the part that
adjusts the particle position.