Restore fluid number for rigid bodies
This fixes an issue with multi-fluid caused by
ef5e40040980f25c83fbf0e5da41dcd5b6f875ee. In practice, each particle may
have either a fluid or object number, or both, so encode the
information in the same field using bitshifts and masks.

For make_particleinfo, the number passed is automatically guessed to be
an object or fluid number depending on particle type, on the assumption
that only rigid bodies or open boundaries need an object number.

Handling of the particle id is also standardized and made independent
from endianness.