XProblem: add support for XYZ files
It is now supported to load a geometry from a XYZ file (point cloud).
It can be loaded as fluid, fixed boundary or moving/floating object;
loading as open boundary does not throw exceptions but is not tested yet.
Similarly to the HDF5 file support, an STL mesh can optionally be given
to enable the geometry with body dynamics.
Loading has been tested with XYZ files produced with Meshlab, with and
without vertex normals.
Shifting according to the origin point is not implemented yet.