XProblem: align object_id assignment with recent policy change in Problem
The object ids are assigned according to the insertion order, but first
the bodies are sorted so that all the forces bodies are at the beginning
(first floating - then moving with feedback - then moving with no
feedback). See Problem::add_moving_body() for details.

Setting the first and last rbindices has also been fixed with the new
policy.