XProblem: fix to commit d769874a (computation of s_hRbFirstIndex[])
"FirstIndex" (which actually is a plain offset) should be computed as

  - first_id_in_object + number_of_previous_body_parts

This was fixed in XProblem, but a regression was introduced in d769874a
while aligning to the new order of bodies.
Since the second addendum is not know until filling is finished,
s_hRbFirstIndex is first filled only with the first_id_in_object, then
fixed in the same cycle where s_hRbLastIndex[] is filled.