XProblem: add callback filterPoints() for erasing filled points
This is called at the end of fill_parts(), right before returning the
number of particles to be allocated. Here we have knowledge about
all the filled points but we are still in time to delete parte of them.

No usage example has been implemented yet, but it is enough to note that
the point vectors are passed by reference.

Note: to delete particles filled by bodies, the specific GeometryInfo
must be retrieved first.