GPUSPH: buildNeibList() at beginning of loop
Rather than having a call outside, and the other inside at the end,
put a single one as first thing in the simulation loop.

The need to put it at the end was driven by a misguided attempt to get
information about new particles when writing, but that is actually
solved by f045301c638315de8f5a8074e1441ac2209e82fd in a more refined
way.