Preliminary exception handling
Handle exceptions being thrown during commands in workers and in
GPUSPH, to get as clean an exit as possible in case of failure.

Workers now set keep_going to false if an exception happens in the
thread, and will avoid carrying on if exceptions happen in other
workers during initialization.

GPUSPH will detect a failed command by keep_going being false
afterwards, and throw an exception from doCommand itself, handling it
at the simulation loop level.