hypot() to compute length accurately
hypot() is the typical name of a function that computes the
hypothenuse of a right triangle without undue loss of accuracy.

This is achieved by taking the largest (in magnitude) of a, b out of
√(a²+b²), thus rewriting the formula as M √(1+(m/M)²) where M =
max(|a|,|b|) and m = min(|a|,|b|). Of course we also want to avoid
producing a NaN when M is zero.

For 3- and 4-vector components a similar approach is used, dividing
the vector by its largest component.