Complete rewrite of MLS
The previous MLS code assembled a matrix that was not
resolution-independent, as the spatial components grew smaller
quadratically with ∆p. In practice this ensured that the matrix was
never (numerically) invertible, and Shepard filtering was used instead,
every time.

We now build a resolution-independent MLS matrix, normalizing the
spatial components by the smoothing length. The normalization has to be
preserved when applying B during the second neighbor loop, so after
solving for B we divide its spatial components by the smoothing length,
again.

We compute an initial guess for B directly, and then refine it using
conjugate residual iterations. When the MLS matrix is invertible, the
initial guess for B is obtained from a perturbed MLS matrix, by adding εΙ
with ε machine epsilon.