Fix StillWater logic for fluid origin
We should only add the extra offset when there are extra layers, which
doesn't happen if we're using DYN_BOUNDARY, but have planes enabled. So
check that dyn_layers is actually larger than 1 before adding the extra
offset, and set dyn_layers to 1 when we are not adding extra layers.

This fixes a bug with all particles ending up in the same ridiculous
position when enabling planes with DYN_BOUNDARY.