Writers: count acting writers and run starting/finishing commands
We introduce a StartWriting command for the Writer control class, which
counts the number of writers that will actually act and call a
start_writing command on each of them. The writers list is then passed
to each writing function (gage, objects etc) and mark_written is called
at the end.

The file counter is now incremented during start_writing, so
next_filenum() is removed.