Make FillIn() safe with negative #layers in Cube, Cylinder, Sphere, Torus
With rects, it makes sense to FillIn() in the opposite side of the normal;
therefore, XProblem calles FillIn() with a negative number of layers.
This works in Rect class but not in the solids, which assume that #layers
is positive (although signed and not checked...).
To make DYN boundaries work while the handling of fills, layers, normals
is redesigned, Cube, Cylinder, Sphere and Torus just invert #layers if
negative.