Surface and normal computation fixes
Since the normal is not scaled, the plane contribution was of the wrong
magnitude, which resulted in particles adjacent to planes being
detected as free surface.

Additionally, the Shepard normalization factor (not used, but saved
with the normals) was being computed incorrectly (using the mass
instead of the volume of the particles).