Add FLUX_COMPUTATION post processing
The new option computes the flux at open boundaries by summing over all
the boundary elements that belong to a certain open boundary. This
required several changes to the post_processing interface such as
passing GlobalData to the routines. Additionally there are two new
functions, one for allocating on the host and one for processing on the
host after the device processing is completed.