Refactor preference for posTex vs posArray
The choice of putting the (read-only) position buffer on a global array
versus a texture is compute-capability dependent.

On devices with a (functioning) L1 cache for global memory, we benefit
from using the array directly because we distribute cache pressure
between L1 and texture caches. On devices without an L1 cache (1.x) it's
better to rely on the texture cache only, so we want to use posTex
instead.

The use of texture vs L1 cache still needs to be tuned on more recent
architectures (due to the L1 cache only being used for register spills;
some tests seem to indicate that at least on Kepler posTex might be
preferrable).

Since adding more sophisticated conditionals would make the code
complicated all around, we define a new constant macro PREFER_L1 which
is set to 1 when posArray should be used, and to 0 when posTex should be
used instead. This macro is then used throughout the code in place of
the explicit check of __COMPUTE__ versus 20. This will make it easier
to add more sohisticated checks on the future.